requirements.txt
accelerate==1.6.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.11
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==75.8.0
six==1.17.0
stack_data==0.6.3
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0